requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
supabase==2.0.3
python-multipart==0.0.6
websockets>=11,<13
